ci: use blacksmith 16vcpu runners
This change switches over the `basic` workflow
to Blacksmith's 16vcpu runners. Additionally, this
change switches the rust cache action to use the
Blacksmith drop-in replacement. The Blacksmith version
of the cache is colocated with Blacksmith runners and is
substantially faster.

The result is that this workflow is overall 2x faster and
half the cost.